backend_fastapi/requirements.txt
# backend_fastapi/requirements.txt
# Core FastAPI
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.2.1

# Pydantic for data validation
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
typing-extensions==4.15.0
annotated-types==0.7.0

# Django for ORM access
django==5.2.10
asgiref==3.11.0
sqlparse==0.5.5

# Database
psycopg2-binary==2.9.11

# Templates
jinja2==3.1.6
markupsafe==3.0.3

# HTTP/Networking
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
httptools==0.7.1
websockets==16.0

# Multipart form handling
python-multipart==0.0.21

# Utilities
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
idna==3.11
tomli==2.4.0
typer==0.21.1
watchfiles==1.1.1
PyJWT

# Tests
pytest==8.0.0
pytest-asyncio==0.23.5
